requirements.txt
apiflask[yaml]
flask-cors
pyinstaller
PyMuPDF
python-dotenv
flask-socketio
SQLAlchemy
openpyxl
